+++ src/components/search/result/CategoryComponent.tsx
