+++ src/components/search/result/ListStudyComponent.tsx
