+++ src/components/search/result/SearchInitComponent.tsx
