+++ src/components/search/result/SearchResultComponent.tsx
